split bacula-sd and bacula-api into separate containers
It is not best practice to have lots of processes all running inside
one container. Additionally, the number of processes that have access
to the storage volume should be as few as possible.

Splitting the bacula-sd into its own container, which only runs one
process, and is the only one with access to the destination storage,
improves the deployment.